apps/optimizer/requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
schedule==1.2.2
pytest==8.3.5
tiktoken==0.9.0
openai==1.76.2
anthropic==0.51.0
